Update to gravity_hpx3.5-fixed race condition, use uint64_t, and use HDF5
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@7005 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ examples/gravity/HDF5_reader.cpp
@@ -32,7 +32,7 @@ vector<point> createvecs (config_f& param) {
  DataSet dataset=file.openDataSet(DATASET_NAME); //Open dataset
  DataSpace dataspace=dataset.getSpace(); //Get the dataspace
  hsize_t dims_out[2];
- int ndims=0;
+ uint64_t ndims=0;
  ndims=dataspace.getSimpleExtentDims(dims_out,NULL); //Get dimentions
 
  //Define a memory block
